+++ NEWS.rst
@@ -6,6 +6,8 @@ every change, see the Git log.
 
 Latest
 ------
+* Minor: Added the check function - a more verbose version of converting the
+  expectation to a boolean value.
 * Patch: Added assert to ensure that expectations has been setup prior to
   evaluation.